feature(DefaultModal): added default modal inside modal service
