+++ src/modals/index.ts
@@ -1,1 +1,1 @@
-export * from './main';
+export * from './main/main';
